+++ src/modules/utils/is-not-equal.ts
@@ -1,0 +1,5 @@
+import * as util from '../utils'
+
+export const isNotEqual = (value: any, valueToCompare: any): boolean => {
+  return !util.isEqual(value, valueToCompare)
+}