+++ src/modules/utils/regex-match.ts
@@ -1,0 +1,3 @@
+export const regexMatch = (value: string, regex: RegExp): boolean => {
+  return regex.test(value)
+}